Implementata la possibilità che dei colpi nemici non vadano a segno (hit chance)

+++ Archer.h
@@ -12,7 +12,7 @@ class Archer : public Enemy {
 public:
     Archer(const sf::Texture &archerTexture, const sf::Texture &bowTexture, const sf::Texture &arrowTexture,
            const sf::Vector2i &spawnTile, const sf::Vector2i &tileSize, const sf::Vector2i &rectSkin,
-           sf::Vector2f damageHit, const std::vector<std::vector<Node>> &nodeMap, bool animated = true, int hp = 15,
+           sf::Vector2f damageHit, const std::vector<std::vector<Node>> &nodeMap, bool animated = true, float hp = 15,
            float speed = 200.f, unsigned int points = 5, unsigned int coins = 1, int armor = 0, bool bubble = false);
 
     void hit(GameCharacter &target) override;
